fix: moved locked scenarios into modules
