+++ packages/wallets/src/features/cashier/flows/WalletFiatOnRamp/WalletFiatOnRamp.tsx
@@ -1,8 +1,7 @@
 import React, { useEffect, useMemo } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useActiveWalletAccount } from '@deriv/api-v2';
-import { FiatOnRampModule } from '../../modules';
-import { CashierLocked } from '../../screens';
+import { CashierLocked, FiatOnRampModule } from '../../modules';
 
 const WalletFiatOnRamp = () => {
     const { data } = useActiveWalletAccount();
