+++ packages/wallets/src/features/cashier/flows/WalletTransfer/WalletTransfer.tsx
@@ -1,8 +1,7 @@
 import React, { useCallback, useEffect } from 'react';
 import { useTransferBetweenAccounts } from '@deriv/api-v2';
 import { Loader } from '../../../../components';
-import { TransferModule } from '../../modules';
-import { CashierLocked } from '../../screens';
+import { CashierLocked, TransferModule } from '../../modules';
 import { TransferNotAvailable } from '../../screens/TransferNotAvailable';
 
 const WalletTransfer = () => {
