+++ packages/wallets/src/features/cashier/modules/CashierLocked/CashierLocked.scss
