+++ packages/wallets/src/features/cashier/modules/CashierLocked/__tests__/CashierLocked.spec.tsx
