+++ packages/wallets/src/features/cashier/modules/DepositLocked/DepositLocked.scss
