+++ packages/wallets/src/features/cashier/modules/WithdrawalLocked/WithdrawalLocked.scss
