+++ packages/wallets/src/features/cashier/modules/index.ts
@@ -1,9 +1,12 @@
+export * from './CashierLocked';
 export * from './DepositCrypto';
 export * from './DepositFiat';
+export * from './DepositLocked';
 export * from './FiatOnRamp';
 export * from './ResetBalance';
 export * from './Transactions';
 export * from './Transfer';
 export * from './WithdrawalCrypto';
 export * from './WithdrawalFiat';
+export * from './WithdrawalLocked';
 export * from './WithdrawalVerification';
