+++ packages/wallets/src/features/cashier/screens/index.ts
@@ -1,4 +1,1 @@
-export * from './CashierLocked';
-export * from './DepositLocked';
 export * from './TransferNotAvailable';
-export * from './WithdrawalLocked';
